// src/unit/enums/unit-messages.enum.ts
export enum UnitSuccessMessages {
  CREATED = "واحد ایجاد شد",
  FOUND = "واحد پیدا شد",
  LISTED = "لیست واحد ها دریافت شد",
  UPDATED = "واحد با موفقیت بروز رسانی شد",
  DELETED = "واحد با موفقیت حذف شد",
}

export enum UnitErrorMessages {
  DUPLICATE = "واحد تکراری است",
  NOT_FOUND = "واحد پیدا نشد",
}